requirements.txt
adbutils             2.7.2
annotated-types      0.7.0
anyio                4.4.0
apkutils2            1.0.0
Appium-Python-Client 4.0.0
attrs                23.2.0
build                1.2.1
CacheControl         0.14.0
cached-property      1.5.2
certifi              2023.11.17
cffi                 1.16.0
charset-normalizer   3.3.2
cigam                0.0.3
cleo                 2.1.0
click                8.1.7
colorama             0.4.6
colorlog             6.7.0
construct            2.10.70
crashtest            0.4.1
decorator            5.1.1
Deprecated           1.2.14
deprecation          2.1.0
distlib              0.3.8
dnspython            2.6.1
dulwich              0.21.7
email_validator      2.1.1
et-xmlfile           1.1.0
exceptiongroup       1.2.1
facebook-wda         1.4.6
fastapi              0.111.0
fastapi-cli          0.0.4
fastjsonschema       2.19.1
filelock             3.13.1
h11                  0.14.0
httpcore             1.0.5
httpretty            1.1.4
httptools            0.6.1
httpx                0.27.0
idna                 3.4
importlib_metadata   7.1.0
installer            0.7.0
jaraco.classes       3.4.0
Jinja2               3.1.4
keyring              24.3.1
logzero              1.7.0
lxml                 4.9.3
markdown-it-py       3.0.0
MarkupSafe           2.1.5
mdurl                0.1.2
more-itertools       10.2.0
msgpack              1.0.8
openpyxl             3.1.2
orjson               3.10.3
outcome              1.3.0.post0
packaging            23.2
pexpect              4.9.0
Pillow               10.1.0
pip                  24.0
pkginfo              1.11.0
platformdirs         4.2.2
poetry               1.8.3
poetry-core          1.9.0
poetry-plugin-export 1.8.0
progress             1.6
ptyprocess           0.7.0
py                   1.11.0
pycparser            2.22
pydantic             2.7.3
pydantic_core        2.18.4
pyelftools           0.30
Pygments             2.18.0
pyproject_hooks      1.1.0
PySocks              1.7.1
python-dotenv        1.0.1
python-multipart     0.0.9
pywin32              306
pywin32-ctypes       0.2.2
PyYAML               6.0.1
rapidfuzz            3.9.3
requests             2.31.0
requests-toolbelt    1.0.0
retry                0.9.2
rich                 13.7.1
selenium             4.21.0
setuptools           70.0.0
shellingham          1.5.4
six                  1.16.0
sniffio              1.3.1
sortedcontainers     2.4.0
starlette            0.37.2
tomli                2.0.1
tomlkit              0.12.5
tornado              6.3.3
trio                 0.25.1
trio-websocket       0.11.1
trove-classifiers    2024.5.22
typer                0.12.3
typing_extensions    4.12.2
uiautodev            0.4.0
uiautomator2         2.16.25
ujson                5.10.0
urllib3              2.1.0
uvicorn              0.30.1
virtualenv           20.26.2
watchfiles           0.22.0
wdapy                0.2.3
websockets           12.0
wheel                0.43.0
whichcraft           0.6.1
wrapt                1.16.0
wsproto              1.2.0
xmltodict            0.13.0
zipp                 3.19.2